+++ src/components/modals/ModalChangeUserData/ModalChangeUserData.module.css
@@ -1,5 +1,5 @@
 .pageContainer {
-  width: 100vw;
+  width: 100%;
   height: 100vh;
   background-color: rgba(0, 0, 0, 0.5);
   position: fixed;
@@ -9,7 +9,7 @@
 }
 
 .modalForm {
-  position: fixed;
+  position: absolute;
   left: 50%;
   top: 50%;
   z-index: 2;
