+++ src/components/profileData/profileData.module.css
@@ -1,0 +1,56 @@
+.title {
+   color: #000000;
+   width: 304px;
+   height: 56px;
+   margin-bottom: 40px;
+   font-size: 48px;
+   line-height: 56px;
+   text-align: left;
+ }
+ 
+ .profile_data {
+   display: flex;
+   flex-direction: column;
+   gap: 20px;
+ }
+ 
+ .login,
+ .password {
+   /* width: 256px; */
+   height: 32px;
+   font-size: 24px;
+   line-height: 32px;
+   text-align: left;
+ }
+ 
+ .button_container {
+   display: flex;
+   flex-direction: column;
+   gap: 14px;
+   margin-top: 20px;
+ }
+ 
+ .change_login_btn,
+ .change_password_btn {
+   width: 275px;
+   height: 52px;
+   border-radius: 46px;
+   border: none;
+   font-size: 18px;
+   line-height: 24px;
+   text-align: center;
+   color: #ffffff;
+   background-color: #580ea2;
+   cursor: pointer;
+ }
+ .modalOverlay {
+   position: fixed;
+   top: 0;
+   left: 0;
+   width: 100%;
+   height: 100%;
+   background-color: rgba(0, 0, 0, 0.5);
+   display: flex;
+   align-items: center;
+   justify-content: center;
+ }
